Move EventTarget::ref/deref to EventTargetInlines.h https://bugs.webkit.org/show_bug.cgi?id=292502
Reviewed by Chris Dumez.

Moved EventTarget::ref/deref from Node.h to EventTargetInlines.h.

* Source/WebCore/Modules/encryptedmedia/legacy/WebKitMediaKeySession.cpp:
* Source/WebCore/dom/EventTarget.h:
* Source/WebCore/dom/EventTargetInlines.h:
(WebCore::EventTarget::ref):
(WebCore::EventTarget::deref):
* Source/WebCore/dom/FocusEvent.cpp:
* Source/WebCore/dom/FocusEvent.h:
* Source/WebCore/dom/MouseEvent.cpp:
(WebCore::MouseEvent::setRelatedTarget):
* Source/WebCore/dom/MouseEvent.h:
* Source/WebCore/dom/MouseEventInit.h:
* Source/WebCore/dom/Node.h:
(WebCore::EventTarget::ref): Deleted.
(WebCore::EventTarget::deref): Deleted.
* Source/WebCore/dom/Touch.cpp:
* Source/WebCore/dom/Touch.h:
* Source/WebCore/dom/WheelEvent.cpp:
* Source/WebCore/dom/ios/PointerEventIOS.cpp:
* Source/WebCore/editing/AlternativeTextController.cpp:
* Source/WebCore/html/HTMLSlotElement.cpp:
* Source/WebCore/html/HTMLSummaryElement.cpp:
* Source/WebCore/inspector/InspectorController.cpp:
* Source/WebCore/inspector/InspectorOverlay.cpp:
* Source/WebCore/rendering/RenderBlock.cpp:

Canonical link: https://commits.webkit.org/294500@main
