+++ Source/WebCore/dom/Touch.cpp
@@ -29,6 +29,7 @@
 
 #include "Touch.h"
 
+#include "EventTargetInlines.h"
 #include "LocalDOMWindow.h"
 #include "LocalFrame.h"
 #include "LocalFrameView.h"
@@ -96,6 +97,8 @@ Touch::Touch(EventTarget* target, int identifier, int clientX, int clientY, int 
     , m_absoluteLocation(absoluteLocation)
 {
 }
+
+Touch::~Touch() = default;
 
 Ref<Touch> Touch::cloneWithNewTarget(EventTarget* eventTarget) const
 {
