+++ Source/WebCore/dom/WheelEvent.cpp
@@ -26,6 +26,7 @@
 
 #include "DataTransfer.h"
 #include "EventNames.h"
+#include "EventTargetInlines.h"
 #include "MouseEventTypes.h"
 #include "Node.h"
 #include <wtf/MathExtras.h>
