+++ Source/WebCore/editing/AlternativeTextController.cpp
@@ -37,6 +37,7 @@
 #include "EditorClient.h"
 #include "Element.h"
 #include "EventLoop.h"
+#include "EventTargetInlines.h"
 #include "FloatQuad.h"
 #include "LocalFrame.h"
 #include "LocalFrameView.h"
