+++ Source/WebCore/html/HTMLSlotElement.cpp
@@ -29,6 +29,7 @@
 #include "AXObjectCache.h"
 #include "ElementInlines.h"
 #include "Event.h"
+#include "EventTargetInlines.h"
 #include "EventNames.h"
 #include "HTMLNames.h"
 #include "InspectorInstrumentation.h"
